requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
openai-whisper>=20231117
laion-clap>=1.1.4
transformers>=4.36.0
pyyaml>=6.0
numpy>=1.24.0
scipy>=1.10.0
